+++ embedded-graphics/src/pixelcolor/mod.rs
@@ -50,9 +50,11 @@
 //!
 //!         Ok(())
 //!     }
+//! }
 //!
+//! impl OriginDimensions for EpdDisplay {
 //!     fn size(&self) -> Size {
-//!         Size::zero()
+//!         Size::new(300, 300)
 //!     }
 //! }
 //!
@@ -76,7 +78,7 @@
 //! draw the specified color and doesn't depend on the raw data conversions,
 //! see the [`raw` module] documentation for an example that uses this feature.
 //!
-//! [`DrawTarget`]: ../trait.DrawTarget.html
+//! [`DrawTarget`]: ../draw_target/trait.DrawTarget.html
 //! [`PixelColor`]: trait.PixelColor.html
 //! [`Raw`]: trait.PixelColor.html#associatedtype.Raw
 //! [`raw` module]: raw/index.html