+++ embedded-graphics/src/prelude.rs
@@ -2,13 +2,14 @@
 
 #[doc(no_inline)]
 pub use crate::{
+    draw_target::{DrawTarget, DrawTargetExt},
     fonts::Font,
-    geometry::{Angle, AngleUnit, Dimensions, Point, Size},
+    geometry::{Angle, AngleUnit, Dimensions, OriginDimensions, Point, Size},
     image::{ImageDimensions, IntoPixelIter},
     pixel_iterator::{IntoPixels, PixelIteratorExt},
     pixelcolor::{raw::RawData, GrayColor, IntoStorage, PixelColor, RgbColor},
     primitives::{ContainsPoint, Primitive},
     style::StyledPrimitiveAreas,
     transform::Transform,
-    DrawTarget, Drawable, Pixel,
+    Drawable, Pixel,
 };